Refactor auth token and support CQ_API_KEY env var
