+++ cli/cmd/login.go
@@ -13,8 +13,8 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/adrg/xdg"
 	"github.com/cenkalti/backoff/v4"
+	"github.com/cloudquery/cloudquery/cli/internal/auth"
 	"github.com/pkg/browser"
 	"github.com/spf13/cobra"
 )
@@ -142,33 +142,12 @@ func runLogin(ctx context.Context) (err error) {
 	if refreshToken == "" {
 		return fmt.Errorf("failed to get refresh token")
 	}
-	err = saveRefreshToken(refreshToken)
+	err = auth.SaveRefreshToken(refreshToken)
 	if err != nil {
 		return fmt.Errorf("failed to save refresh token: %w", err)
 	}
 
 	fmt.Println("CLI successfully authenticated.")
 
-	return nil
-}
-
-func saveRefreshToken(refreshToken string) error {
-	tokenFilePath, err := xdg.DataFile("cloudquery/token")
-	if err != nil {
-		return fmt.Errorf("can't determine a proper location for token file: %w", err)
-	}
-	tokenFile, err := os.OpenFile(tokenFilePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0o644)
-	if err != nil {
-		return fmt.Errorf("can't open token file %q for writing: %w", tokenFilePath, err)
-	}
-	defer func() {
-		e := tokenFile.Close()
-		if err == nil && e != nil {
-			err = fmt.Errorf("can't close token file %q after writing: %w", tokenFilePath, e)
-		}
-	}()
-	if _, err = tokenFile.WriteString(refreshToken); err != nil {
-		return fmt.Errorf("failed to write token to %q: %w", tokenFilePath, err)
-	}
 	return nil
 }
