+++ cli/cmd/testdata/cloudquery/token
@@ -1,0 +1,1 @@
+testtoken
